+++ packages/ui/src/Rate/RateStars/RateStars.stories.tsx
@@ -2,68 +2,64 @@ import React, {useLayoutEffect, useRef} from 'react';
 import {Story, Meta} from '@storybook/react';
 
 import RateStars, {RateStarsProps, MARKS_ARRAY} from './';
-import TableGuides, {
-  TableGuidesProps,
-} from '../../../../../.storybook/TableGuides';
 
-const RATE_SIZES: NonNullable<RateStarsProps['size']>[] = ['s', 'm', 'l'];
+const RATE_SIZES: number[] = [51, 30, 20];
 
 export default {
   title: 'RateStars',
-} as Meta;
-
-type RateStoryMeta = Omit<
-  TableGuidesProps<Omit<RateStarsProps, ''>>,
-  'Component'
-> & {
-  name: string;
-};
-
-const rateStoryMeta: RateStoryMeta = {
-  name: 'Text field',
-  cols: RATE_SIZES.map((size) => ({
-    key: size.toLocaleUpperCase(),
-    props: {size},
-  })),
-  rows: [...MARKS_ARRAY, '5+'].map((mark) => ({
-    key: mark,
-    props: {value: mark},
-  })),
-};
+} as Meta<RateStarsProps>;
 
 const Template: Story<RateStarsProps & {hovered?: boolean}> = ({
   hovered,
   ...args
-}) => (
-  <TableGuides
-    cols={rateStoryMeta.cols}
-    rows={rateStoryMeta.rows}
-    Component={(props) => {
-      const ref = useRef<HTMLDivElement | null>(null);
-      useLayoutEffect(() => {
-        if (!hovered) return;
-        const event = new MouseEvent('mouseover', {
-          bubbles: true,
-          cancelable: true,
-        });
+}) => {
+  const ref = useRef<HTMLDivElement | null>(null);
+  useLayoutEffect(() => {
+    if (!hovered) return;
+    const event = new MouseEvent('mouseover', {
+      bubbles: true,
+      cancelable: true,
+    });
 
-        ref.current.dispatchEvent(event);
-      }, [hovered]);
+    ref.current.dispatchEvent(event);
+  }, [hovered]);
 
-      return <RateStars ref={ref} {...args} {...props} />;
-    }}
-  />
-);
+  return (
+    <div style={{display: 'flex', flexDirection: 'row'}}>
+      {RATE_SIZES.map((size) => (
+        <div style={{display: 'flex', flexDirection: 'column'}}>
+          {[...MARKS_ARRAY, '5+'].map((mark) => (
+            <RateStars
+              ref={ref}
+              value={mark}
+              style={{fontSize: `${size}px`}}
+              {...args}
+            />
+          ))}
+        </div>
+      ))}
+    </div>
+  );
+};
 
 export const Default = Template.bind({});
 Default.storyName = 'Default';
-Default.args = {
-  onChange: undefined,
-};
+Default.args = {};
 
 export const Hover = Template.bind({});
 Hover.storyName = 'Hover';
 Hover.args = {
-  onChange: undefined,
   hovered: true,
+};
+
+export const Stroked = Template.bind({});
+Stroked.storyName = 'Stroked';
+Stroked.args = {
+  design: 'stroked',
+};
+
+export const Light = Template.bind({});
+Light.storyName = 'Light';
+Light.args = {
+  design: 'light',
 };