+++ packages/ui/src/Rate/RateTooltip/index.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, {forwardRef, HTMLAttributes} from 'react';
+import cx from 'classnames';
 
 import Tooltip from '../../Tooltip';
 import {Text} from '../../Typography';
@@ -9,33 +10,40 @@ import styles from './RateTooltip.module.scss';
 
 type Props = {
   show: boolean;
-  size: 's' | 'm' | 'l';
-  children: React.ReactNode;
   trigger?: ITrigger;
-};
+} & Omit<HTMLAttributes<HTMLDivElement>, 'onChange'>;
 
-export default function RateTooltip({
-  show,
-  size,
-  children,
-  trigger = 'hover',
-}: Props): JSX.Element {
-  if (show) {
-    return (
-      <Tooltip persist trigger={trigger} className={styles['tooltip']}>
-        <Tooltip.Content
-          className={styles['tooltipContent']}
-          overlayClassName={styles['tooltipOverlay']}
-          position="top-center"
+export const RateTooltip = forwardRef<HTMLDivElement, Props>(
+  ({show, children, className, trigger = 'hover', ...props}, ref) => {
+    if (show) {
+      return (
+        <Tooltip
+          persist
+          trigger={trigger}
+          className={cx(styles['tooltip'], className)}
+          {...props}
+          ref={ref}
         >
-          <Text size={size} bold>
-            Пять&nbsp;с&nbsp;плюсом
-          </Text>
-        </Tooltip.Content>
-        <Tooltip.Toggler>{children}</Tooltip.Toggler>
-      </Tooltip>
-    );
-  }
+          <Tooltip.Content
+            className={styles['tooltipContent']}
+            overlayClassName={styles['tooltipOverlay']}
+            position="top-center"
+          >
+            <Text size="m" bold>
+              Пять&nbsp;с&nbsp;плюсом
+            </Text>
+          </Tooltip.Content>
+          <Tooltip.Toggler>{children}</Tooltip.Toggler>
+        </Tooltip>
+      );
+    }
 
-  return <>{children}</>;
-}
+    return (
+      <div className={className} {...props} ref={ref}>
+        {children}
+      </div>
+    );
+  },
+);
+
+export default RateTooltip;